+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * Test flight for turn-into restrictions in edit-mode and multi-select mode (#376)
 
+### Design & UX 🔳
+
+* Apple emojis or uploaded image in block-action-toolbar link's preview (#630)
+
 ### Fixes & tech 🚒
 
 * Should not show toast when clicking on markup url (#698)